refactor: add signal builder
